# api/openapi.yaml
components:
  schemas:
...
    ErrorResponse:
      additionalProperties: true
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
      type: object
...
    ExploreCacheUnavailableResponse:
      additionalProperties: true
      properties:
        error:
          type: string
        message:
          type: string
        readiness:
          enum:
            - absent
            - interrupted
            - stale_schema
            - drifted
          type: string
        recovery_action:
          type: string
      required:
        - error
        - message
        - readiness
        - recovery_action
      type: object
...
    ExploreHTTPRequest:
      additionalProperties: false
      properties:
        candidate_snapshot_id:
          type: string
        cursor:
          type: string
        filters:
          items:
            $ref: "#/components/schemas/ExploreFilter"
          type:
            - array
            - "null"
        grouping:
          items:
            $ref: "#/components/schemas/ExploreGroupDimension"
          type:
            - array
            - "null"
        limit:
          format: int64
          maximum: 500
          minimum: 0
          type: integer
        presentation:
          enum:
            - table
            - timeline
            - files
          type: string
        query:
          type: string
        search_mode:
          enum:
            - full_text
            - semantic
            - hybrid
          type: string
        sort:
          items:
            $ref: "#/components/schemas/ExploreSort"
          type:
            - array
            - "null"
      type: object
    ExploreHTTPResponse:
      additionalProperties: true
      properties:
        cache_revision:
          type: string
        candidate_pool_saturated:
          type: boolean
        candidate_snapshot_id:
          type: string
        next_cursor:
          type: string
        rows:
          items:
            $ref: "#/components/schemas/EntryRow"
          type: array
        search_provenance:
          $ref: "#/components/schemas/SearchProvenance"
        total_count:
          format: int64
          type: integer
      required:
        - rows
        - cache_revision
        - search_provenance
      type: object
...
    FileGroupsHTTPRequest:
      additionalProperties: false
      properties:
        cursor:
          type: string
        filename_query:
          type: string
        grouping:
          items:
            $ref: "#/components/schemas/ExploreGroupDimension"
          maxItems: 1
          minItems: 1
          type: array
        limit:
          format: int64
          maximum: 500
          minimum: 0
          type: integer
        mime_families:
          items:
            type: string
          type:
            - array
            - "null"
        predicate:
          $ref: "#/components/schemas/ExploreHTTPRequest"
        sort:
          items:
            $ref: "#/components/schemas/ExploreGroupSort"
          maxItems: 1
          type:
            - array
            - "null"
      required:
        - predicate
        - grouping
      type: object
    FileGroupsHTTPResponse:
      additionalProperties: true
      properties:
        cache_revision:
          type: string
        candidate_snapshot_id:
          type: string
        next_cursor:
          type: string
        rows:
          items:
            $ref: "#/components/schemas/ExploreGroupRow"
          type: array
        search_provenance:
          $ref: "#/components/schemas/SearchProvenance"
        total_count:
          format: int64
          type: integer
      required:
        - rows
        - total_count
        - cache_revision
        - search_provenance
      type: object
    FileMetadataResponse:
      additionalProperties: true
      properties:
        content_available:
          type: boolean
        content_hash:
          type: string
        content_state:
          enum:
            - metadata_only
            - url_only
            - missing_blob
            - local_content
          type: string
        conversation_id:
          format: int64
          type: integer
        filename:
          type: string
        id:
          format: int64
          type: integer
        message_id:
          format: int64
          type: integer
        mime_type:
          type: string
        size_bytes:
          format: int64
          type: integer
        url:
          type: string
      required:
        - id
        - message_id
        - conversation_id
        - filename
        - mime_type
        - size_bytes
        - content_state
        - content_available
      type: object
    FileSearchHTTPRequest:
      additionalProperties: false
      properties:
        cursor:
          type: string
        filename_query:
          type: string
        limit:
          format: int64
          maximum: 500
          minimum: 0
          type: integer
        mime_families:
          items:
            type: string
          type:
            - array
            - "null"
        predicate:
          $ref: "#/components/schemas/ExploreHTTPRequest"
        sort:
          $ref: "#/components/schemas/FileSearchSort"
      required:
        - predicate
        - sort
      type: object
    FileSearchHTTPResponse:
      additionalProperties: true
      properties:
        cache_revision:
          type: string
        candidate_snapshot_id:
          type: string
        files:
          items:
            $ref: "#/components/schemas/FileSearchRow"
          type:
            - array
            - "null"
        next_cursor:
          type: string
        search_provenance:
          $ref: "#/components/schemas/SearchProvenance"
        total_count:
          format: int64
          type: integer
      required:
        - files
        - total_count
        - cache_revision
        - search_provenance
      type: object
    FileSearchRow:
      additionalProperties: true
      properties:
        containing_title:
          type: string
        content_available:
          type: boolean
        content_state:
          enum:
            - metadata_only
            - url_only
            - missing_blob
            - local_content
          type: string
        conversation_id:
          format: int64
          type: integer
        entry_key:
          type: string
        filename:
          type: string
        id:
          format: int64
          type: integer
        key:
          type: string
        message_id:
          format: int64
          type: integer
        mime_family:
          type: string
        mime_type:
          type: string
        occurred_at:
          format: date-time
          type: string
        participant_domains:
          items:
            type: string
          type:
            - array
            - "null"
        participant_ids:
          items:
            format: int64
            type: integer
          type:
            - array
            - "null"
        participant_labels:
          items:
            type: string
          type:
            - array
            - "null"
        size_bytes:
          format: int64
          type: integer
        source_id:
          format: int64
          type: integer
        source_identifier:
          type: string
        source_type:
          type: string
      required:
        - id
        - key
        - entry_key
        - message_id
        - conversation_id
        - occurred_at
        - source_id
        - source_type
        - source_identifier
        - containing_title
        - filename
        - mime_type
        - mime_family
        - size_bytes
        - content_state
        - content_available
      type: object
    FileSearchSort:
      additionalProperties: false
      properties:
        direction:
          type: string
        field:
          type: string
      required:
        - field
        - direction
      type: object
...
  securitySchemes:
    apiKey:
      in: header
      name: X-Api-Key
      type: apiKey
info:
  title: msgvault API
  version: 1.24.0
openapi: 3.1.0
paths:
...
  /api/v1/domains/{domain}/files/search:
    post:
      operationId: searchDomainFiles
      parameters:
        - description: Exact normalized domain fact
          in: path
          name: domain
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileSearchHTTPRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileSearchHTTPResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "503":
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/ExploreCacheUnavailableResponse"
                  - $ref: "#/components/schemas/ErrorResponse"
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Search one domain's analytical files
      tags:
        - Exploration
...
  /api/v1/explore:
    post:
      operationId: explore
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExploreHTTPRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExploreHTTPResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "503":
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/ExploreCacheUnavailableResponse"
                  - $ref: "#/components/schemas/ErrorResponse"
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Explore canonical archive entries
      tags:
        - Exploration
...
  /api/v1/files/groups:
    post:
      operationId: groupFiles
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileGroupsHTTPRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileGroupsHTTPResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "503":
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/ExploreCacheUnavailableResponse"
                  - $ref: "#/components/schemas/ErrorResponse"
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Group analytical files
      tags:
        - Exploration
  /api/v1/files/search:
    post:
      operationId: searchFiles
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileSearchHTTPRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileSearchHTTPResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "503":
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/ExploreCacheUnavailableResponse"
                  - $ref: "#/components/schemas/ErrorResponse"
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Search analytical files
      tags:
        - Exploration
  /api/v1/files/{id}:
    get:
      operationId: getFile
      parameters:
        - description: File attachment ID
          in: path
          name: id
          required: true
          schema:
            format: int64
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileMetadataResponse"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Get authoritative file metadata
      tags:
        - API
...
  /api/v1/people/{id}/files/search:
    post:
      operationId: searchPersonFiles
      parameters:
        - description: Durable participant ID
          in: path
          name: id
          required: true
          schema:
            format: int64
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileSearchHTTPRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileSearchHTTPResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "503":
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/ExploreCacheUnavailableResponse"
                  - $ref: "#/components/schemas/ErrorResponse"
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Search one person's analytical files
      tags:
        - Exploration
...